move temp dir to own utils, exports it to environment
