+++ agent.py
@@ -44,8 +44,7 @@ from config import (
     load_check,
     generate_jmx_configs,
     _is_affirmative,
-    SD_PIPE_NAME,
-    setup_temp_dir
+    SD_PIPE_NAME
 )
 from daemon import AgentSupervisor, Daemon
 from emitter import http_emitter
@@ -64,6 +63,7 @@ from utils.service_discovery.config_stores import get_config_store
 from utils.service_discovery.sd_backend import get_sd_backend
 from utils.watchdog import Watchdog
 from utils.windows_configuration import get_sdk_integration_paths
+from utils.tmp_files import setup_temp_dir
 
 # Constants
 PID_NAME = "dd-agent"
@@ -473,7 +473,7 @@ def main():
     hostname = get_hostname(agentConfig)
     in_developer_mode = agentConfig.get('developer_mode')
 
-    setup_temp_dir(agentConfig, proc_name='agent')
+    setup_temp_dir(agentConfig, proc_name='collector')
 
     COMMANDS_AGENT = [
         'start',
